added error handling to server routes

+++ app/api/server_routes.py
@@ -20,11 +20,14 @@ def servers():
 @server_routes.route('/user')
 @login_required
 def users_servers():
-    # get list of all the server memberships to the current user
+     # get list of all the server memberships to the current user
     user_memberships = current_user.server_memberships
-    # looping through user_memberships to get that membership.server
-    servers = [membership.server for membership in user_memberships]
-    return {"servers": [server.to_dict() for server in servers]}, 200
+    if (len(user_memberships) > 0):
+        # looping through user_memberships to get that membership.server
+        servers = [membership.server for membership in user_memberships]
+        return {"servers": [server.to_dict() for server in servers]}, 200
+    else:
+        return {"errors": ["User has no servers"]}
 
 # CREATE SERVER
 @server_routes.route("/", methods=['POST'])
@@ -42,6 +45,7 @@ def create_server():
         db.session.add(server)
         db.session.commit()
         return {"server": server.to_dict()}, 201
+    return {"errors": ["Could not complete request"]}
 
 # UPDATE SERVER
 @server_routes.route("/<int:id>", methods=['PUT'])
@@ -50,6 +54,9 @@ def update_server(id):
     form = ServerForm()
     form["csrf_token"].data = request.cookies["csrf_token"]
     server = Server.query.get(id)
+    userId = int(current_user.id)
+    if (userId != server.owner_id):
+        return {"errors": ["Unauthorized"]}, 400
     if server:
         server.name = form.data["name"]
         server.icon_url = form.data["icon_url"]
@@ -58,11 +65,15 @@ def update_server(id):
         db.session.commit()
         return {"server": server.to_dict()}
 
+
 # DELETE SERVER
 @server_routes.route("<int:id>", methods=['DELETE'])
 @login_required
 def delete_server(id):
     server = Server.query.get(id)
+    userId = int(current_user.id)
+    if (userId != server.owner_id):
+        return {"errors": ["Unauthorized"]}, 400
     if server:
         db.session.delete(server)
         db.session.commit()
